+++ saleor/payment/gateways/adyen/tests/conftest.py
@@ -76,3 +76,56 @@ def payment_adyen_for_order(payment_adyen_for_checkout, order_with_lines):
     payment_adyen_for_checkout.order = order_with_lines
     payment_adyen_for_checkout.save()
     return payment_adyen_for_checkout
+
+
+@pytest.fixture()
+def notification():
+    def fun(
+        event_code=None,
+        success=None,
+        psp_reference=None,
+        merchant_reference=None,
+        value=None,
+    ):
+        event_code = event_code or "AUTHORISATION"
+        success = success or "true"
+        psp_reference = psp_reference or "852595499936560C"
+        merchant_reference = merchant_reference or "UGF5bWVudDoxNw=="
+        value = value or 1130
+
+        return {
+            "additionalData": {},
+            "eventCode": event_code,
+            "success": success,
+            "eventDate": "2019-06-28T18:03:50+01:00",
+            "merchantAccountCode": "SaleorECOM",
+            "pspReference": psp_reference,
+            "merchantReference": merchant_reference,
+            "amount": {"value": value, "currency": "USD"},
+        }
+
+    return fun
+
+
+@pytest.fixture
+def notification_with_hmac_signature():
+    return {
+        "additionalData": {
+            "expiryDate": "12/2012",
+            " NAME1 ": "VALUE1",
+            "cardSummary": "7777",
+            "totalFraudScore": "10",
+            "hmacSignature": "D4bKVtjx5AlBL2eeQZIh1p7G1Lh6vWjzwkDlzC+PoMo=",
+            "NAME2": "  VALUE2  ",
+            "fraudCheck-6-ShopperIpUsage": "10",
+        },
+        "amount": {"currency": "GBP", "value": 20150},
+        "eventCode": "AUTHORISATION",
+        "eventDate": "2020-07-24T12:40:22+02:00",
+        "merchantAccountCode": "SaleorPOS",
+        "merchantReference": "8313842560770001",
+        "paymentMethod": "visa",
+        "pspReference": "test_AUTHORISATION_4",
+        "reason": "REFUSED",
+        "success": "false",
+    }
