+++ saleor/payment/gateways/adyen/tests/webhooks/cassettes/test_handle_authorization_with_auto_capture.yaml
